fix: San 的类型文件的一些修复
